Move LOI feature generation to LoiRepository
