+++ ground/src/main/java/com/google/android/ground/persistence/local/room/stores/RoomSubmissionStore.kt
@@ -220,13 +220,13 @@ class RoomSubmissionStore @Inject internal constructor() : LocalSubmissionStore 
     submissionMutationDao.loadAllOnceAndStream()
 
   override suspend fun findByLocationOfInterestId(
-    id: String,
+    loidId: String,
     vararg states: MutationEntitySyncStatus
-  ): List<SubmissionMutationEntity> = submissionMutationDao.findByLocationOfInterestId(id, *states)
+  ): List<SubmissionMutationEntity> = submissionMutationDao.findByLocationOfInterestId(loidId, *states)
 
-  override suspend fun getPendingSubmissionCountByLocationOfInterestId(id: String): Int =
-    submissionMutationDao.getPendingSubmissionCountByLocationOfInterestId(
-      id,
+  override suspend fun getPendingSubmissionCountByLocationOfInterestId(loiId: String): Int =
+    submissionMutationDao.getSubmissionMutationCount(
+      loiId,
       MutationEntityType.CREATE,
       MutationEntitySyncStatus.PENDING
     )
