+++ ground/src/main/java/com/google/android/ground/persistence/local/stores/LocalSubmissionStore.kt
@@ -55,7 +55,7 @@ interface LocalSubmissionStore : LocalMutationStore<SubmissionMutation, Submissi
   fun getAllMutationsAndStream(): Flowable<List<SubmissionMutationEntity>>
 
   suspend fun findByLocationOfInterestId(
-    id: String,
+    loidId: String,
     vararg states: MutationEntitySyncStatus
   ): List<SubmissionMutationEntity>
